% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DX_init.R
\name{dx_init}
\alias{dx_init}
\title{DX_init}
\usage{
dx_init(K = 47, S = 1)
}
\arguments{
\item{K}{number of random numbers wanted}

\item{S}{number of random numbers wanted}
}
\value{

}
\description{
DX_init
}
\examples{
dx_init()

}
